fix bug

+++ src/components/SavingPlan.jsx
@@ -5,30 +5,43 @@ import LocaleContext from '../context/LocaleContext';
 import { deleteSavings } from '../utils/savings';
 import UserContext from '../context/UserContext';
 import SavingPlanItemLoading from './SavingPlanItemLoading';
+import nodata from '../assets/images/no-data.svg';
 
 function SavingPlan({ savings, loading, refresh }) {
   const { locale } = React.useContext(LocaleContext);
   const { user } = React.useContext(UserContext);
 
   const onDeleteHandler = async (id) => {
-    Swal.showLoading();
-    const deleteHandler = await deleteSavings(id, user.uid);
-    if (deleteHandler.success) {
-      Swal.fire({
-        icon: 'success',
-        title: `${locale === 'en' ? 'Delete Saving Plan Success' : 'Berhasil Menghapus Rencana Tabungan'}`,
-        showConfirmButton: false,
-        timer: 1500,
-      });
-      refresh();
-    } else {
-      Swal.fire({
-        icon: 'error',
-        title: deleteHandler.message,
-        showConfirmButton: false,
-        timer: 1500,
-      });
-    }
+    Swal.fire({
+      title: `${locale === 'en' ? 'Delete This Saving Plan?' : 'Hapus Rencana Tabungan?'}`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#013496',
+      cancelButtonColor: '#DC3545',
+      confirmButtonText: `${locale === 'en' ? 'Delete' : 'Hapus'}`,
+      cancelButtonText: `${locale === 'en' ? 'Cancel' : 'Batalkan'}`,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        Swal.showLoading();
+        const deleteHandler = await deleteSavings(id, user.uid);
+        if (deleteHandler.success) {
+          Swal.fire({
+            icon: 'success',
+            title: `${locale === 'en' ? 'Delete Saving Plan Success' : 'Berhasil Menghapus Rencana Tabungan'}`,
+            showConfirmButton: false,
+            timer: 1500,
+          });
+          refresh();
+        } else {
+          Swal.fire({
+            icon: 'error',
+            title: deleteHandler.message,
+            showConfirmButton: false,
+            timer: 1500,
+          });
+        }
+      }
+    });
   };
 
   if (loading) {
@@ -50,9 +63,10 @@ function SavingPlan({ savings, loading, refresh }) {
   }
 
   return (
-    <h4 className="text-center fw-bold medium__font">
-      {locale === 'en' ? "You don't have savings item" : 'Anda tidak memiliki item tabungan'}
-    </h4>
+    <div className="empty-data mt-5 pt-4">
+      <img className="empty-data__image d-block mx-auto" src={nodata} alt="no data" />
+      <h2 className="medium__font text-center mt-3 fw-bold">{locale === 'en' ? "You Don't Have Savings Item" : 'Anda Tidak Memiliki Item Tabungan'}</h2>
+    </div>
   );
 }
